+++ code/src/App.js
@@ -2,8 +2,9 @@ import React, { useState } from 'react'
 import { BrowserRouter, Switch, Route } from 'react-router-dom'
 
 import Header from './components/Header'
-import MovieList from './components/MovieList'
+import MovieList from './pages/MovieList'
 import MovieInfo from './pages/MovieInfo'
+import CompanyPage from 'pages/CompanyPage'
 
 export const App = () => {
   const [movieListType, setMovieListType] = useState('popular')
@@ -22,6 +23,11 @@ export const App = () => {
         <Route path="/movies/:id">
           <MovieInfo />
         </Route>
+
+        <Route path="/companies/:id">
+          <CompanyPage />
+        </Route>
+
       </Switch>
     </BrowserRouter>
   )
