+++ code/src/pages/CompanyPage.js
@@ -1,0 +1,11 @@
+import React, {useEffect, useState} from 'react'
+
+import {useParams} from 'react-router-dom'
+
+const CompanyPage = () => {
+  return (
+    <div>Company</div>
+  )
+}
+
+export default CompanyPage
